requirements.txt
aws-cdk-lib==2.56.0
constructs==10.1.*